Journal clue update functionality added
Needs to be linked with puzzle interaction but that's a 10 second job.

Also reshuffled the entire DoorPuzzle.cs so that it pulls the answer out of GameStateManager.cs - that way, we have a much more obvious, centralised location for game data like the level solution to be kept.

Added a couple of tags as well in preparation for linking puzzle tiles into journal updates and interaction with the player.